fix(KEN-801): the whole package page follows an update started on it
The Projects tab's Update and Update all commit through the updates store,
which refreshes the scan and the audit and knows nothing about the page's own
three reads. The card's install date was made to follow the landed commit in
#1799; the Overview kept showing the files and version of the copy the update
replaced, and the header kept offering an update already applied.

The page's reads are keyed on the commit installed in the place it names, the
same signal the Projects tab re-reads on, so a landed update reads them again
and an unrelated store touch reads nothing.

Two absorbed issues ride in the same file. KEN-946: a failed record or
timeline read no longer collapses to a bare null. Each carries a ReadState,
and a failure is worded through the note slot the header already has, so a
page that withholds Update says which read did not land instead of naming a
check nobody needs. KEN-806: each Projects card's Update and Remove carry the
place in their accessible name, so two cards' buttons no longer read alike.

The file-size ratchet took two files over their class thresholds. The version-
changing actions move out of use-package-data.ts, which now only reads, and
the package page's test fixtures move to src/test/package-page.tsx, shared by
package.test.tsx and the new package-update.test.tsx the way the account
store's harness already is.
