+++ changelog.d/fixed/KEN-801.md
@@ -1,0 +1,3 @@
+- Updating a package from its page's Projects tab now refreshes the whole
+  page, the tab's buttons name the place they act on, and a read that failed
+  says so where Update would be.
