+++ ui/src/components/package/use-package-data.ts
@@ -5,28 +5,23 @@ import {
   type PackageDiff,
   type PackageFile,
   type PackageMeta_Serialize,
-  type PackageUpdate_Serialize,
   type Scope,
   type VersionRow,
 } from "@/bindings";
+import { packageReadFailedNote } from "@/lib/copy-updates";
+import { installedCommits } from "@/lib/package-places";
 import {
-  FOLLOW_SOURCE_TOAST,
-  updatedToastLabel,
-  VERSION_ERROR_TITLE,
-} from "@/lib/copy";
-import { UPDATES_ONE_AT_A_TIME_NOTE } from "@/lib/copy-updates";
+  READ_LANDED,
+  READ_PENDING,
+  type ReadState,
+  readOf,
+} from "@/lib/read-state";
 import { sameScope } from "@/lib/scope";
-import { settled } from "@/lib/settled";
-import { saying } from "@/lib/undone";
-import { workOut } from "@/lib/updates-read-state";
-import { versionRowLabel } from "@/lib/versions";
 import { useAuditStore } from "@/stores/audit";
 import { useEditorStore } from "@/stores/editor";
 import type { PackageRef } from "@/stores/nav";
 import { useProblemsStore } from "@/stores/problems";
-import { useScanStore } from "@/stores/scan";
-import { holdingBusy, useUpdatesStore } from "@/stores/updates";
-import { sayApply } from "@/stores/updates-apply";
+import { useUpdatesStore } from "@/stores/updates";
 
 export type PackageView =
   | { mode: "files"; file: string | null }
@@ -50,11 +45,56 @@ export const diffHarness = (
 ): HarnessId | null =>
   view.mode === "diff" && view.harness ? view.harness : primary;
 
-/** The package page's reads, refetchable as one unit after a mutation. */
-export function usePackageData(ref: PackageRef | null) {
+/** The two reads Update rests on, as one answer: the record that says held
+ *  or following, and the timeline it moves along. Failed the moment either
+ *  did, because a gate nobody could read is one nothing can speak for, and
+ *  landed only once both have. The file list is not in here: no Update ever
+ *  turned on it, and folding it in would withhold the button over a read it
+ *  does not depend on. */
+const bothReads = (meta: ReadState, versions: ReadState): ReadState => {
+  if (meta.status === "failed") return meta;
+  if (versions.status === "failed") return versions;
+  return meta.status === "landed" && versions.status === "landed"
+    ? READ_LANDED
+    : READ_PENDING;
+};
+
+/** Why the package page has no Update when its own reads are the reason,
+ *  or null when they are not. Silent while they are pending: the page is
+ *  still filling in, and a header note on every open is noise rather than
+ *  news. */
+export const packageReadNote = ({ status, error }: ReadState): string | null =>
+  status === "failed" && error !== null ? packageReadFailedNote(error) : null;
+
+/** The package page's reads, refetchable as one unit after a mutation.
+ *
+ *  Read again when the commit installed in the place this page names moves:
+ *  an update started from the Projects tab commits through the updates
+ *  store and never through `load`, so without that the Overview would go on
+ *  showing the files and version of the copy the update replaced and the
+ *  header would go on offering an update already applied. Keyed on the
+ *  commit, the way the Projects tab's own re-read is, so an unrelated
+ *  updates-store touch reads nothing. */
+export function usePackageData(ref: PackageRef | null): {
+  meta: PackageMeta_Serialize | null;
+  files: PackageFile[];
+  versions: VersionRow[];
+  read: ReadState;
+  load: () => void;
+} {
   const [meta, setMeta] = useState<PackageMeta_Serialize | null>(null);
   const [files, setFiles] = useState<PackageFile[]>([]);
   const [versions, setVersions] = useState<VersionRow[]>([]);
+  // How each of the two went, kept beside the values: a read that failed
+  // leaves the same empty page as one that found nothing, and the reason it
+  // came back with is the only thing that tells them apart.
+  const [metaRead, setMetaRead] = useState<ReadState>(READ_PENDING);
+  const [versionsRead, setVersionsRead] = useState<ReadState>(READ_PENDING);
+  const commit = useUpdatesStore((s) =>
+    ref === null
+      ? ""
+      : installedCommits(s.rows, ref.kind, ref.name, [ref.scope]),
+  );
 
   const load = useCallback(() => {
     if (!ref) return;
@@ -62,6 +102,7 @@ export function usePackageData(ref: PackageRef | null) {
       .packageMeta(ref.scope, ref.kind, ref.name)
       .then((response) => {
         setMeta(response.status === "ok" ? response.data : null);
+        setMetaRead(readOf(response));
       });
     void commands
       .packageFiles(ref.scope, ref.kind, ref.name)
@@ -72,11 +113,19 @@ export function usePackageData(ref: PackageRef | null) {
       .packageVersions(ref.scope, ref.kind, ref.name)
       .then((response) => {
         setVersions(response.status === "ok" ? response.data : []);
+        setVersionsRead(readOf(response));
       });
   }, [ref]);
 
-  useEffect(load, [load]);
-  return { meta, files, versions, load };
+  // biome-ignore lint/correctness/useExhaustiveDependencies: the commit a landed update moves, not a value `load` closes over
+  useEffect(load, [load, commit]);
+  return {
+    meta,
+    files,
+    versions,
+    read: bothReads(metaRead, versionsRead),
+    load,
+  };
 }
 
 /** The diff behind a diff view, fetched when the view asks for one. The
@@ -120,95 +169,6 @@ export function usePackageDiff(
   }, [ref, view, harness, showError]);
 
   return diff;
-}
-
-/** The version-changing actions for one package, each applying the whole
- *  scope and refreshing the app's derived state after. `setBusy` drives
- *  the page's spinner; `reload` refetches the package's own data. */
-export function packageVersionActions(
-  ref: PackageRef,
-  displayName: string,
-  held: boolean,
-  setBusy: (busy: boolean) => void,
-  reload: () => void,
-) {
-  const showError = (message: string) =>
-    useProblemsStore
-      .getState()
-      .showError({ title: VERSION_ERROR_TITLE, message });
-  const afterChange = () => {
-    reload();
-    void useScanStore.getState().refresh();
-    void useAuditStore.getState().refresh({ force: true });
-  };
-  // Every one of these applies a plan that can refuse a rendering, so
-  // none of them toasts off the click: the command's own report says what
-  // reached the files, and `done` is only this surface's word for having
-  // written the package. This page has no edited-row filter, and a refusal
-  // is broader than an edit anyway — files kendex never put there, a
-  // provenance clash — so the held answer arrives here whatever the page
-  // believes about edits.
-  // Under the updates store's `busy` as well as the page's own spinner —
-  // these commit like any update does, and one write at a time is what the
-  // Updates page's check and its own writes both rest on.
-  const run = async (
-    call: () => Promise<
-      | { status: "ok"; data: PackageUpdate_Serialize }
-      | { status: "error"; error: string }
-    >,
-    done: string,
-  ) => {
-    // Asked before the command is sent: navigating here mid-write is the
-    // overlap that flag rules out.
-    if (workOut(useUpdatesStore.getState()))
-      return showError(UPDATES_ONE_AT_A_TIME_NOTE);
-    setBusy(true);
-    return holdingBusy(async () => {
-      // A transport failure rejects rather than refusing. Unwrapped it
-      // would skip the report, leave `setBusy` up for the life of the view
-      // and skip the read-back this promises either way. `saying` because
-      // a hold move can take a declaring package away with it.
-      const response = saying(await settled(call()));
-      setBusy(false);
-      if (response.status === "error") {
-        showError(response.error);
-        // An error is not proof that nothing changed: `package_set_rev`
-        // persists the revision and only then applies, so a failed apply
-        // answers over a manifest that already moved. The page reads back
-        // either way, or it shows the old version as settled.
-        afterChange();
-        return;
-      }
-      // One package's apply, so a removal it reports is that package's.
-      sayApply(done, response.data, 1);
-      afterChange();
-    });
-  };
-
-  const switchTo = (row: VersionRow) =>
-    run(
-      () => commands.packageSetRev(ref.scope, ref.kind, ref.name, row.id),
-      updatedToastLabel(`${displayName} to ${versionRowLabel(row)}`),
-    );
-
-  // A held package moves its hold to the latest; a follower is brought
-  // current by the single-package apply — Update never silently pins a
-  // follower, and does not move the scope's other followers along.
-  const updateToLatest = (latest: VersionRow) =>
-    held
-      ? switchTo(latest)
-      : run(
-          () => commands.packageUpdate(ref.scope, ref.kind, ref.name),
-          updatedToastLabel(displayName),
-        );
-
-  const follow = () =>
-    run(
-      () => commands.packageSetRev(ref.scope, ref.kind, ref.name, null),
-      FOLLOW_SOURCE_TOAST,
-    );
-
-  return { switchTo, updateToLatest, follow };
 }
 
 /** One gate for every control that rewrites this package's manifest: the
