+++ ui/src/lib/versions.ts
@@ -26,14 +26,17 @@ export function installedRow(rows: VersionRow[]): VersionRow | undefined {
  *  off its version rows, not off the update row, so the timeline it draws
  *  and the button above it agree.
  *
- *  `withheld` is the reason, not a verdict — `updates-read-state.ts`
- *  [`packageUpdateNote`], which answers the read's own state and delegates
- *  the rest to [`updateWithheld`] — and it is the whole of the update
- *  read's say here. There is deliberately no second gate on how that read
- *  went: a reason that cannot be said cannot hide the button, and a page
- *  that refuses without saying why is the shape that costs. So this asks
- *  only what the page itself knows — a version to move to and from, and
- *  the meta read that says held or following — and defers the rest. */
+ *  `withheld` is the reason, not a verdict: the first reason either read
+ *  behind the page carries, the page's own (`use-package-data.ts`
+ *  [`packageReadNote`]) ranking ahead of the update read's
+ *  (`updates-read-state.ts` [`packageUpdateNote`], which answers that
+ *  read's state and delegates the rest to [`updateWithheld`]). Every gate
+ *  past the two below reaches this through that one string, which is the
+ *  rule and not a habit of the caller: a reason that cannot be said cannot
+ *  hide the button, and a page that refuses without saying why is the shape
+ *  that costs. So this asks only what the page itself knows — a version to
+ *  move to and from, and the meta read that says held or following — and
+ *  defers the rest. */
 export const canUpdatePackage = (page: {
   latest: VersionRow | undefined;
   installed: VersionRow | undefined;
